Release/0.2.0

+++ src/main/ipc/ipc.ts
@@ -1,18 +1,36 @@
-import { app, ipcMain } from 'electron';
-import { FingerprintGenerator } from 'fingerprint-generator';
+import { ipcMain } from 'electron';
 import { ProfileManager } from 'main/browser-profile/profile-manager';
 import Store from 'electron-store';
 import path from 'path';
-import { spawn } from 'child_process';
-import { openSync } from 'fs';
-import { orderBy } from 'lodash';
-import { ManageBrowserProfileDto } from 'shared/models/renderer-data-schema';
-import { MainResponse } from 'shared/ipc';
+import { ManageBrowserProfileDto, BrowserStatusDto, BrowserStatus } from 'shared/models';
+import { MainResponse, Channel } from 'shared/ipc';
+import { Chromium } from 'main/chromium';
+import { rmSync } from 'fs';
 import { IState } from '../state/istate';
-import Channel from './channel';
 
-const startIpc = (state: IState): void => {
+export const registerIpcHandlers = (state: IState): void => {
   const profileManager = new ProfileManager(state.store);
+
+  const getProfileDir = (profileId: string) => {
+    return path.join(path.dirname(state.store.path), 'browser-profiles', profileId);
+  };
+
+  const deleteProfileDir = (profileId: string) => {
+    const browserProfilePath = getProfileDir(profileId);
+    rmSync(browserProfilePath, { recursive: true, force: true, maxRetries: 5 });
+  };
+
+  const browserStatusChanged = (status: BrowserStatusDto) => {
+    if (state.mainWindow === null) return;
+
+    state.mainWindow.webContents.send(Channel.ProfileStatusChange, status);
+  };
+
+  ipcMain.handle(Channel.CloseApp, async () => {
+    state.killAllBrowserWindows();
+    state.appCloseConfirmed = true;
+    state.mainWindow?.close();
+  });
 
   ipcMain.handle(Channel.SaveProfile, async (event, profile: ManageBrowserProfileDto) => {
     let savedProfile = null;
@@ -30,7 +48,14 @@ const startIpc = (state: IState): void => {
   });
 
   ipcMain.handle(Channel.DeleteProfile, async (event, id: string) => {
+    if (state.activeBrowserWindows.has(id)) {
+      return MainResponse.error(
+        new Error('Profile is being used. Stop the session first')
+      );
+    }
+
     try {
+      deleteProfileDir(id);
       const removed = profileManager.delete(id);
       return MainResponse.success(removed);
     } catch (error: any) {
@@ -38,57 +63,101 @@ const startIpc = (state: IState): void => {
     }
   });
 
-  ipcMain.handle(Channel.LaunchProfile, (event, id: string) => {
-    const allProfiles = state.store?.get('profiles') || [];
-    const selectedProfile = allProfiles.find((p) => p.id === id);
-    if (selectedProfile === undefined) throw new Error();
+  ipcMain.handle(Channel.StartBrowser, async (event, id: string) => {
+    if (state.activeBrowserWindows.has(id)) {
+      return MainResponse.error(
+        new Error('Browser with this profile is already running')
+      );
+    }
 
-    const generator = new FingerprintGenerator({
-      operatingSystems: [selectedProfile.os],
-      browsers: ['chrome'],
-      devices: ['desktop'],
-    });
-    const fingerprintWHeaders = generator.getFingerprint();
-    const profilePath = path.join(
-      app.getPath('userData'),
-      'pbrowser',
-      'browser-profile',
-      id
-    );
-    const fingerprintStore = new Store({
-      cwd: profilePath,
-      name: 'fingerprint',
-    });
-    fingerprintStore.set('fingerprint', fingerprintWHeaders.fingerprint);
-    const outFile = openSync(`${profilePath}/chrome-out.log`, 'a');
-    const errFile = openSync(`${profilePath}/chrome-err.log`, 'a');
-    const browserLocation = 'E:\\Downloads\\chromium\\chrome-win\\chrome.exe';
-    const args = [`--user-data-dir=${profilePath}`];
-    const childProc = spawn(browserLocation, args, {
-      stdio: ['ignore', outFile, errFile],
-    });
-    console.log(`PID is: ${childProc.pid}`);
+    try {
+      browserStatusChanged(new BrowserStatusDto(id, BrowserStatus.PendingActive));
+      profileManager.updateLaunchDate(id);
 
-    childProc.on('close', (code) => {
-      console.log(`child process close with code ${code}`);
-    });
+      const profile = profileManager.get(id);
+      const browserProfilePath = getProfileDir(id);
+      const fingerprintStore = new Store({
+        cwd: browserProfilePath,
+        name: 'fingerprint',
+      });
+      fingerprintStore.set('fingerprint', profile);
 
-    childProc.on('error', (err) => {
-      console.log(`child process exited with error ${err}`);
-      console.log(err);
-    });
+      const instance = new Chromium(id, browserProfilePath);
+      state.activeBrowserWindows.set(id, instance);
+      instance.start();
+
+      instance.process?.once('spawn', () => {
+        browserStatusChanged(new BrowserStatusDto(id, BrowserStatus.Active));
+      });
+      instance.process?.once('exit', () => {
+        browserStatusChanged(new BrowserStatusDto(id, BrowserStatus.PendingInactive));
+      });
+      instance.process?.once('close', (code) => {
+        state.activeBrowserWindows.delete(id);
+        // set Inactive status only if closed normally or killed, otherwise set Error
+        if (code === 0 || code === 1 || code === null) {
+          browserStatusChanged(new BrowserStatusDto(id, BrowserStatus.Inactive));
+        }
+      });
+      instance.process?.once('error', () => {
+        browserStatusChanged(new BrowserStatusDto(id, BrowserStatus.StartError));
+      });
+
+      return MainResponse.success();
+    } catch (error: any) {
+      state.activeBrowserWindows.delete(id);
+      browserStatusChanged(new BrowserStatusDto(id, BrowserStatus.StartError));
+      return MainResponse.error(error);
+    }
+  });
+
+  ipcMain.handle(Channel.StopBrowser, async (event, id: string) => {
+    if (!state.activeBrowserWindows.has(id))
+      return MainResponse.error(new Error('Browser with this profile is not running'));
+
+    try {
+      const instance = state.activeBrowserWindows.get(id);
+      instance?.kill();
+
+      return MainResponse.success();
+    } catch (error: any) {
+      browserStatusChanged(new BrowserStatusDto(id, BrowserStatus.StopError));
+      return MainResponse.error(error);
+    }
   });
 
   ipcMain.handle(Channel.GetProfiles, async () => {
     const allProfiles = profileManager.getAll();
-    const orderedProfiles = orderBy(
-      allProfiles,
-      [(p) => p.lastLaunchDate || '1970-01-01T00:00:00.000Z', (p) => p.name],
-      ['desc', 'asc']
-    );
 
-    return orderedProfiles;
+    return allProfiles;
+  });
+
+  ipcMain.handle(Channel.GetActiveBrowserWindows, async () => {
+    return Array.from(state.activeBrowserWindows.keys()).map((id) => ({
+      profileId: id,
+      status: BrowserStatus.Active,
+    }));
+  });
+
+  ipcMain.handle(Channel.DeleteSession, async (event, profileId: string) => {
+    if (state.activeBrowserWindows.has(profileId)) {
+      return MainResponse.error(
+        new Error('Profile is being used. Stop the session first')
+      );
+    }
+
+    try {
+      deleteProfileDir(profileId);
+
+      return MainResponse.success();
+    } catch (error: any) {
+      return MainResponse.error(error);
+    }
   });
 };
 
-export default startIpc;
+export const deregisterIpcHandlers = () => {
+  Object.values(Channel).forEach((channel) => {
+    ipcMain.removeHandler(channel);
+  });
+};
